delete and refactor  legacy task and dataloader endpoint and tests
